Fixes issue with custom settings file
When the buildClient variables are contained in a custom settings file,
then the custom variables are not yet available in the plugin constructor.
Presumably, the serverless modules first initializes the plugin and only
then parses the files referenced in the config file.

This fix moves the setting of the configuration variable to the "beforeClientBuild" hook.